+++ src/imio/annex/content/events.py
@@ -9,9 +9,10 @@ Created by mpeeters
 
 from collective.documentviewer.settings import GlobalSettings
 from collective.documentviewer.utils import allowedDocumentType
+from collective.iconifiedcategory.utils import get_category_object
+from collective.iconifiedcategory.utils import update_categorized_elements
 from plone import api
 from zope.event import notify
-from zope.lifecycleevent import ObjectModifiedEvent
 
 from imio.annex.events import AnnexFileChangedEvent
 from imio.annex.content.annex import IAnnex
@@ -44,7 +45,8 @@ def annex_conversion_started(obj, event):
     container = obj.aq_parent
     if obj.UID() not in getattr(container, 'categorized_elements', {}):
         return
-    notify(ObjectModifiedEvent(obj))
+    target = get_category_object(obj, obj.content_category)
+    update_categorized_elements(container, obj, target)
 
 
 def annex_conversion_finished(obj, event):
@@ -53,4 +55,5 @@ def annex_conversion_finished(obj, event):
     container = obj.aq_parent
     if obj.UID() not in getattr(container, 'categorized_elements', {}):
         return
-    notify(ObjectModifiedEvent(obj))
+    target = get_category_object(obj, obj.content_category)
+    update_categorized_elements(container, obj, target)